// resources/js/game/shop/shop-table/enums/shop-item-types.ts
export enum ShopItemTypes {

    // Weapon Types
    WEAPON = 'weapon',
    STAVE = 'stave',
    BOW = 'bow',
    HAMMER = 'hammer',
    GUN = 'gun',
    FAN = 'fan',
    SCRATCH_AWL = 'scratch-awl',
    MACE = 'mace',

    // Armour Types
    BODY = 'body',
    SLEEVES = 'sleeves',
    GLOVES = 'gloves',
    BOOTS = 'boots',
    SHIELD = 'shield',
    HELMET = 'helmet',
    LEGGINGS = 'leggings',
}